lp-reserves-feeder/requirements.txt
python-dotenv
web3_multicall
snowflake-connector-python
confluent-kafka
colorlog
